Split into C and headers
