+++ ut/serial_ut_r.cc
@@ -1,0 +1,28 @@
+#include "gtest/gtest.h" 
+#include "gmock/gmock.h"
+#include "Arduino.h"
+#include "executor.h"
+#include "comm.h"
+#include "comm_defs.h"
+#include "comm_common.h"
+#include "serial.h"
+#include "temp_measurement.h"
+#include "mock.h"
+
+using ::testing::AtLeast;  
+using ::testing::Return;
+using ::testing::InSequence;
+using ::testing::NiceMock;
+using ::testing::StrictMock;
+using ::testing::_;
+
+
+
+
+
+TEST_F(StrictSerialFixture, check_RTV)
+{
+    
+    
+}
+
